requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.7
google-generativeai>=0.3.1
sqlalchemy>=2.0.0
pandas>=2.0.0
flask>=2.0.0
python-dotenv>=1.0.0